pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "signature-capture"
version = "0.1.0"
dependencies = [
    "pyserial>=3.5",
    "Pillow>=9.0",
]
requires-python = ">=3.8"
authors = [
    {name = "Jhordy", email = "jhordygg17@gmail.com"},
]
description = "Una librería para capturar y procesar firmas desde un Microcontrolador"
readme = "README.md"
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]